% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recessions.R
\name{add_recessions}
\alias{add_recessions}
\title{Add recessions to time series graphs}
\usage{
add_recessions(min = 2000, max = 2010, xdata = c("int", "date"), text
  = TRUE, text_nudge_x = .2, fill = "#002d49", alpha = 0.11)
}
\arguments{
\item{min}{Numeric, the year of the earliest recession to display.}

\item{max}{Numeric, the year of the latest recession to display.}

\item{xdata}{Char, a string indicating whether the x axis of the primary data
being graphed is in integer or date format. This argument will accept any
of the following values: \code{c("int", "integer", "integers", "i", "date",
"dates", "d")}}

\item{text}{Logical, whether or not to include labels that identify each box
as a recession.}

\item{text_nudge_x}{Numeric, the amount to shift the labels along the x axis.
Measured as a distance from the end of each recession (a value of 0 will
place the label on the edge of each rectangle). The default of 0.2 seems to
work well in many situations but may need to be overridden.}

\item{fill}{Char, the color of the rectangles to draw.}

\item{alpha}{Numeric, the opacity of the rectangles to draw.}
}
\description{
\code{add_recessions} returns ggplot geoms that add rectangles representing
recessions to a plot. It will either return a single \code{geom_rect} geom or
\code{geom_rect} and \code{geom_text} geoms, where the text places the word
"recession" near each rect.
}
\section{Default color}{
 The CMAP color palette gray used for recessions is
  \code{#e3e8eb}. The default fill and alpha values of \code{#002d49} and
  \code{0.11} replicate the palette color at the highest possible
  transparency. This is done because there is no known way to place the
  recession geom behind the graph's grid lines. The default therefore
  produces the approved CMAP color while altering the appearance of any
  overlapping grid lines as little as possible. This was generated using the
  hints found here:
  \url{https://stackoverflow.com/questions/6672374/convert-rgb-to-rgba-over-white}.
}

\examples{
grp_goods <- dplyr::filter(grp_over_time, category == "Goods-Producing")
grp_goods <- dplyr::mutate(grp_goods, year2 = as.Date(lubridate::date_decimal(year)))

# INTEGER X AXIS:
ggplot(data = grp_goods, mapping = aes(x = year, y = realgrp, color = cluster)) +
  add_recessions(min = 2007, max = 2019, xdata = "int", text = TRUE) +
  geom_line() +
  scale_x_continuous("Year") +
  theme_minimal()

# DATE X AXIS:
ggplot(data = grp_goods,
mapping = aes(x = year2, y = realgrp, color = cluster)) +
  add_recessions(min = 2007, max = 2019, xdata = "date", text = TRUE, text_nudge_x = 100) +
  geom_line() +
  scale_x_date("Year") +
  theme_minimal()

# BELOW EXAMPLES SHOW MORE THAN 1 RECESSION
\dontrun{
library(tidyverse)

time_series <- tibble(date_dec = 1800:2020, var = rnorm(221), var2 = rnorm(221)) \%>\%
pivot_longer(cols = starts_with("var"),
             names_to = "var") \%>\%
  mutate(date_date = as.Date(lubridate::date_decimal(date_dec))) \%>\%
  select(date_dec, date_date, var, value)

# A plot with an integer-based x axis
ggplot(data = filter(time_series, date_dec >= 1980 & date_dec <= 2019),
       mapping = aes(x = date_dec, y = value, color = var)) +
  add_recessions(min = 1980, max = 2019, xdata = "int", text = TRUE) +
  geom_line() +
  scale_x_continuous("Year") +
  theme_cmap()

# A plot with a date-based x axis
ggplot(data = filter(time_series, date_dec >= 1980 & date_dec <= 2019),
       mapping = aes(x = date_date, y = value, color = var)) +
  add_recessions(min = 1980, max = 2019, xdata = "date", text = TRUE) +
  geom_line() +
  scale_x_date() +
  theme_cmap()
}

}
\seealso{
\url{https://htmlpreview.github.io/?https://github.com/brodieG/ggbg/blob/development/inst/doc/extensions.html#stat-compute}
\url{https://rpubs.com/hadley/97970}
\url{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}
}
