% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_aes.R
\name{calc_cmap_default_aes}
\alias{calc_cmap_default_aes}
\alias{apply_cmap_default_aes}
\alias{unapply_cmap_default_aes}
\title{Calculate, fetch, and set aesthetic defaults}
\usage{
calc_cmap_default_aes(quietly = FALSE)

apply_cmap_default_aes(quietly = FALSE)

unapply_cmap_default_aes(quietly = FALSE)
}
\arguments{
\item{quietly}{Should the function suppress confirmation messages?}
}
\description{
Aesthetics for ggplot2's "geoms" (e.g. the default font for a label or width
for a line) cannot be set via a theme. However, ggplot2 allows these defaults
to be overridden at the package level. These functions work together to align
the most commonly used geoms to CMAP's design style. They are rarely needed
by users except in specific circumstances.
}
\details{
Because ggplot2 does not allow for the "theming" of aesthetics, applying
\code{theme_cmap()} will change attributes of the axes, legends, etc, but not
the attributes of geoms. \code{finalize_plot()} implicitly makes these
changes for the user. However, you may want plots to use these defaults
outside of \code{finalize_plot()}. Run \code{apply_cmap_default_aes()} to use CMAP
defaults for all ggplot2 plots drawn in the current session. To reset to
\code{ggplot2} defaults (technically, to whatever the defaults were when
\code{cmapplot} was loaded), use \code{unapply_cmap_default_aes}.

\code{calc_cmap_default_aes} interprets \code{cmapplot_globals} to figure out
what CMAP default aesthetics should be. It is run automatically when the cmapplot
package is loaded. Manual overrides to \code{cmapplot_globals}, which can be
triggered by \code{set_cmapplot_global()}, will not automatically change
these defaults. Run this after modifying global variables if you want your
changes to impact geom aesthetics.

The geoms these functions currently override are listed in
\code{cmapplot_globals$geoms_to_change}.
}
\section{Functions}{
\itemize{
\item \code{calc_cmap_default_aes}: Calculate CMAP default aesthetics

\item \code{apply_cmap_default_aes}: Apply CMAP aesthetic defaults to all ggplots in
the current session

\item \code{unapply_cmap_default_aes}: Reset modified geom aesthetics to their values
when \code{cmapplot} was first loaded
}}

\examples{

# IMPROVE AND BROADEN TO ACCOUNT FOR NEW FUNCTIONS

\dontrun{
g <- ggplot(filter(grp_over_time, category == "Services"),
            aes(x = year, y = realgrp, color = cluster)) +
  geom_recessions(ymax = 0.4, text_nudge_x = 0.1) +
  theme_cmap(hline = 0,
             axislines = "x",
             legend.max.columns = 2) +
  ggtitle("Change in gross regional product over time") +
  geom_line() +
  scale_x_continuous("Year", breaks = seq(2007, 2017, 2)) +
  coord_cartesian(clip = "off") +
geom_text_lastonly(aes(label = realgrp), add_points = TRUE)

# print normally
g

# overwrite `default_aes`
apply_cmap_default_aes()
g

# reset `default_aes`
unapply_cmap_default_aes()
g

# finalize alters the defaults implicitly, but then resets them automatically.
finalize_plot(g)

# you can also use `finalize` without these modifications
finalize_plot(g, use_cmap_aes = FALSE)
}


}
