% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_discrete.R
\name{cmap_fill_discrete}
\alias{cmap_fill_discrete}
\alias{cmap_color_discrete}
\alias{cmap_colour_discrete}
\title{Apply discrete CMAP palettes to ggplot2 aesthetics}
\usage{
cmap_fill_discrete(palette = "prosperity", reverse = FALSE, ...)

cmap_color_discrete(palette = "prosperity", reverse = FALSE, ...)

cmap_colour_discrete(palette = "prosperity", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose from 'cmap_palettes' list, or use one of the gradients
defined in the 'cmap_gradients' list (gradients will be automatically
converted into discrete bins)}

\item{reverse}{Logical; reverse color order?}

\item{...}{Additional parameters passed on to the scale type}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or
color). See \code{link{cmap_palettes}} for a listing of available gradients.
}
\section{Functions}{
\itemize{
\item \code{cmap_fill_discrete()}: For fill aesthetic

\item \code{cmap_color_discrete()}: For color aesthetic

\item \code{cmap_colour_discrete()}: For color aesthetic

}}
\examples{
ggplot(pop_and_laborforce_by_age, aes(x = variable, y = value, fill = age)) +
   geom_col(position = position_stack(reverse = TRUE)) +
   facet_wrap(~year) +
   cmap_fill_discrete(palette = "community")

ggplot(percentile_wages, aes(x = percentile, y = wage, color = cluster)) +
   geom_line() +
   cmap_color_discrete(palette = "prosperity")

}
