% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_race.R
\name{cmap_fill_race}
\alias{cmap_fill_race}
\alias{cmap_color_race}
\alias{cmap_colour_race}
\title{Apply official CMAP race/ethnicity chart colors to ggplot2 aesthetics}
\usage{
cmap_fill_race(white, black, hispanic, asian, other)

cmap_color_race(white, black, hispanic, asian, other)

cmap_colour_race(white, black, hispanic, asian, other)
}
\arguments{
\item{white}{White/Caucasian}

\item{black}{Black/African American}

\item{hispanic}{Hispanic/Latino}

\item{asian}{Asian}

\item{other}{Other}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or color).
Specify your dataset's unique race factor names (as case-sensitive strings) in the arguments.
All categories are optional in case your dataset does not have some of them.
}
\section{Functions}{
\itemize{
\item \code{cmap_fill_race}: For fill aesthetic

\item \code{cmap_color_race}: For color aesthetic

\item \code{cmap_colour_race}: For color aesthetic
}}

\examples{
ggplot(dplyr::filter(traded_emp_by_race, Race!="Regional average" &
                     variable=="SpecializedTraded")) +
   geom_col(aes(x = Race, y = value, fill = Race)) +
   cmap_fill_race(white = "White", black = "Black",
                  hispanic = "Hispanic", asian = "Asian",
                  other = "Other")

}
