% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_geom_defaults.R
\name{cmap_geom_defaults}
\alias{cmap_geom_defaults}
\alias{apply_cmap_geom_defaults}
\alias{unapply_cmap_geom_defaults}
\title{Fetch and set aesthetic defaults}
\usage{
apply_cmap_geom_defaults(quietly = FALSE)

unapply_cmap_geom_defaults(quietly = FALSE)
}
\arguments{
\item{quietly}{Should the function suppress confirmation message?}
}
\description{
These functions allow for fetching and setting default aesthetic values for
certain ggplot2 geoms. This is necessary for geoms to be "themed" to CMAP
style standards, because (at least at the moment) setting geom aesthetic
defaults on a plot-by-plot basis (such as with \code{ggplot2::theme}) is not
possible. The geoms impacted are stored in
\code{cmapplot_globals$geom_changes}, and at the moment include \code{Line},
\code{Text}, and the text-based custom \code{cmapplot} geoms \code{TextLast}
and \code{RecessionsText}.
}
\details{
These functions are employed implictly within \code{\link{finalize_plot}} to
apply preferred aesthetic defaults to final outputs. They are only necessary
to use explicitly if you would like plots to use these defaults
pre-\code{finalize}.

CAUTION: Running \code{apply_cmap_geom_defaults} will set defaults for all
ggplot2 plots drawn in the current session. To reset to \code{ggplot2}
defaults (technically, to whatever the defaults were when \code{cmapplot}
was loaded), use \code{unapply_cmap_geom_defaults}.

Note: CMAP aesthetic values are hardwired into \code{apply_cmap_geom_defaults}.
}
\section{Functions}{
\itemize{
\item \code{apply_cmap_geom_defaults}: Apply CMAP-preferred defaults to select geoms
for the current session

\item \code{unapply_cmap_geom_defaults}: Reset modified geom aesthetics to their values
when \code{cmapplot} was first loaded
}}

\examples{
\dontrun{
g <- ggplot(mtcars, aes(x = mpg, y = disp, label = rownames(mtcars))) +
  geom_line() +
  geom_text() +
  geom_text_lastonly() +
  theme_minimal() +
  coord_cartesian(clip = "off")

# print normally
g

# overwite defaults
apply_cmap_geom_defaults()
g

# reset to original defaults
unapply_cmap_geom_defaults()
g

# finalize alters the defaults implicitly, but then resets them automatically.
finalize_plot(g)

# you can also use `finalize` without these modifications
finalize_plot(g, use_cmap_aes = FALSE)
}

}
