% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_continuous.R
\name{cmap_pal_continuous}
\alias{cmap_pal_continuous}
\title{Continuous palette prep function}
\usage{
cmap_pal_continuous(palette = "red_gradient", reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'cmap_gradients' list}

\item{reverse}{Logical; reverse color order?}
}
\description{
Continuous palette prep function
}
