% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_discrete.R
\name{cmap_pal_discrete}
\alias{cmap_pal_discrete}
\title{Discrete palette prep function}
\usage{
cmap_pal_discrete(palette = "prosperity", reverse = FALSE)
}
\description{
Discrete palette prep function
}
