% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_discrete.R
\docType{data}
\name{cmap_palettes}
\alias{cmap_palettes}
\alias{viz_palette}
\title{CMAP discrete color palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
cmap_palettes

viz_palette(pal, ttl = deparse(substitute(pal)), num = length(pal))
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colours representing a palette}

\item{ttl}{character, title to be displayed (the name of the palette)}

\item{num}{numeric, the number of colours to display}
}
\description{
A sample of Comms-approved colors arranged into palettes

Visualize function is directly from the \href{https://github.com/ropenscilabs/ochRe}{ochRe package}
}
\section{Functions}{
\itemize{
\item \code{viz_palette}: function to vizualize a palette
}}

\examples{
# Run the following function to visualize each discrete palette
purrr::walk2(cmap_palettes, names(cmap_palettes), viz_palette)

}
\keyword{datasets}
