% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\docType{data}
\name{cmapplot_globals}
\alias{cmapplot_globals}
\title{cmapplot global variables}
\format{
An object of class \code{list} of length 5.
}
\usage{
cmapplot_globals
}
\description{
Expose a list of predefined variables for use by the cmapplot package and its
users.
}
\keyword{datasets}
