% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pandemics.R, R/geom_recessions.R,
%   R/geom_text_lastonly.R, R/geom_text_lastonly_repel.R
\name{customproto}
\alias{customproto}
\alias{GeomPandemics}
\alias{GeomPandemicsText}
\alias{GeomRecessions}
\alias{GeomRecessionsText}
\alias{GeomTextLast}
\alias{GeomPointLast}
\alias{GeomTextLastRepel}
\title{Custom ggproto classes}
\description{
The \code{cmapplot} package contains a few custom ggproto objects. For the
most part, these are slightly tweaked versions of ggplot2's default proto
objects. For more information about these, see
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}.

The \code{cmapplot} package contains a few custom ggproto objects. For the
most part, these are slightly tweaked versions of ggplot2's default proto
objects. For more information about these, see
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::ggplot2-ggproto}}.
}
\section{Functions}{
\itemize{
\item \code{GeomPandemics}: Add Pandemic bars to plot.

\item \code{GeomPandemicsText}: Add Pandemic bar labels to plot.

\item \code{GeomRecessions}: Add recession bars to plot.

\item \code{GeomRecessionsText}: Add recession bar labels to plot.

\item \code{GeomTextLast}: Add text to plot for maximum x-value in dataset only.

\item \code{GeomPointLast}: Add points to plot for maximum x-value in dataset only.

\item \code{GeomTextLastRepel}: Add text to plot for maximum x-value in dataset only.

\item \code{GeomPointLast}: Add points to plot for maximum x-value in dataset only.

}}
