% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\docType{data}
\name{.bigpt}
\alias{.bigpt}
\title{Establish `.bigpt` constant}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.bigpt
}
\description{
Divide a size expressed in bigpoints (1/72 inch) by this constant to get the
size of text in millimeters, which is often the size ggplot expects when a
unit object is not accepted.
}
\details{
The numerator, 72, is bigpts/inch. The denominator, 25.4, is mm/inch. E.g. 72
bigpts / 72 bigpts/inch * 25.4 mm/inch  = 25.4 mm. This constant is similar
to `ggplot2::.pt`, but the variation is needed here because this package
consistently expresses units in bigpts.
}
\seealso{
https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size
}
\keyword{datasets}
