% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.lwd}
\alias{.lwd}
\alias{gg_lwd_convert}
\title{Line width conversion}
\usage{
.lwd

gg_lwd_convert(value, unit = "bigpts")
}
\arguments{
\item{value}{Numeric, the value to be converted.}

\item{unit}{Char, the unit of the value to be converted. Can be any of the
units accepted by \code{grid::unit()}, including "bigpts", "pt", "mm", and
"in". Default is \code{bigpts}.}
}
\description{
The factor \code{.lwd} is used to calculate correct output sizes for line
widths. For line widths in \code{ggplot2}, the size in mm must be divided
by this factor for correct output. Because the user is likely to prefer
other units besides for mm, \code{gg_lwd_convert()} is provided as a
convenience function, converting from any unit all the way to ggplot units.
}
\details{
\code{.lwd} is equal to \code{ggplot2::.stroke / ggplot2::.pt}. In
\code{ggplot2}, the size in mm is divided by \code{.lwd} to achieve the
correct output. In the \code{grid} package, however, the size in points
(\code{pts} (or maybe \code{bigpts}? Unclear.) must be divided by
\code{.lwd}. The user is unlikely to interact directly with \code{grid},
but this is how \code{finalize_plot()} does its work.

This is closely related to \code{ggplot::.pt}, which is the factor that
font sizes (in \code{pts}) must be divided by for text geoms within
\code{ggplot2}. Confusingly, \code{.pt} is not required for \code{ggplot2}
font sizes outside the plot area: e.g. axis titles, etc.
}
\section{Functions}{
\itemize{
\item \code{gg_lwd_convert()}: Function to convert from any unit directly to ggplot2's
preferred millimeters.

}}
\examples{
ggplot() + coord_cartesian(xlim = c(-3, 3), ylim = c(-3, 3)) +

  # a green line 3 points wide
  geom_hline(yintercept = 1, color = "green", size = gg_lwd_convert(3)) +

  # black text of size 24 points
  annotate("text", -2, 0, label = "text", size = 24/ggplot2::.pt)


# a blue line 6 points wide, drawn over the plot with  the `grid` package
grid::grid.lines(y = 0.4, gp = grid::gpar(col = "blue", lwd = 6 / .lwd))

}
\seealso{
grid's \code{\link[grid]{unit}}, ggplot2's
  \code{\link[ggplot2]{.pt}}, and
  \url{https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size}
}
