% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_plot2.R
\name{finalize_plot2}
\alias{finalize_plot2}
\title{Arrange and save CMAP ggplot chart}
\usage{
finalize_plot2(
  plot = ggplot2::last_plot(),
  title = "Title here",
  subtitle = "Subtitle here",
  mode = c("plot", "newwindow", "object", "svg", "png", "tiff", "pdf"),
  width = 670,
  height = 400,
  title_width = 150,
  filepath = NULL,
  plot_margins = grid::unit(c(5, 5, 5, 10), "bigpts")
)
}
\arguments{
\item{plot}{ggplot object, the variable name of the plot you have created that
you want to finalize. The efault is \code{ggplot2::last_plot()}, so if
unspecified, the most recent plot will be used.}

\item{title}{Char, the text you want to appear in the title block.}

\item{subtitle}{Char, the text you want to appear in the subtitle block.}

\item{mode}{Char, the action to be taken with the plot. Default is to view in
the window (using "plot"). Other options include "newwindow" (displays in a
new window), "object" (exports as grob object), or one of four file
extensions that can be saved (png, tiff, svg, and pdf).}

\item{width}{Numeric, the width in pixels for the image, including the title.
Default = 670.}

\item{height}{Numeric, the height in pixels for the image. Default = 400.}

\item{title_width}{Numeric, the width in pixels for the title. Default = 150.}

\item{filepath}{Char, the filepath you want the plot to be saved to.  You can
specify an extension to use, or an extension will be added if you specify it
in "mode".}

\item{plot_margins}{Vector of units, the margins around the elements of the
plot within the plot object. This requires a vector of 4 "unit" elements,
defining the margins clockwise starting from the top. The default is 5, 5,
5, and 10 big points on the top, right, bottom, and left, respectively.
Inputs should be formatted as
grid::unit(c(`top`,`right`,`bottom`,`left`),"`units`")}
}
\description{
'\code{finalize_plot2} will save a ggplot into a frame defined by  CMAP design
standards. It will align your title and subtitle to the left, add a horizontal
line on top, and make other adjustments. If instructed to do so, it will also
save the plot in one of four file formats (png, tiff, svg, and pdf) to a
specified location. This function will not apply CMAP design standards to the
plot itself--use \code{theme_cmap()} for that.
}
\examples{
\dontrun{
econ_plot <-
  cluster_jobchange \%>\%
  ggplot(aes(x = reorder(name, jobchange), y = jobchange, fill = category)) +
  theme_cmap(gridlines = "v", hline = 0) +
  geom_col() +
  coord_flip() +
  scale_y_continuous(labels = scales::comma)

econ_title <- "Change in employment in specified clusters in the Chicago
  Metropolitan Statistical Area, 2001-17."
econ_subtitle <- "Source: Chicago
  Metropolitan Agency for Planning analysis of traded clusters."

finalize_plot2(econ_plot,
               econ_title,
               econ_subtitle,
               mode = "plot",
               plot_margins = unit(c(5,20,5,10),"bigpts"))

transit_plot <- transit_ridership \%>\%
  mutate(system = case_when(
    system == "cta_bus" ~ "CTA (Bus)",
    system == "cta_rail" ~ "CTA (Rail)",
    system == "metra" ~ "Metra",
    system == "pace" ~ "Pace",
    system == "pace_ada" ~ "Paratransit"
  )) \%>\%
  ggplot(aes(x = year, y = ridership, color = system)) +
  geom_line() +
  theme_cmap()

transit_title <- "Transit ridership in the RTA region over time, 1980-2019
  (in millions)"
transit_subtitle <- "Source: Chicago Metropolitan Agency for Planning
  analysis of data from the Regional Transportation Authority."

finalize_plot2(transit_plot,
               transit_title,
               transit_subtitle,
               mode="pdf",
               filepath = "foo",
               height = 300,
               title_width = 200,
               width = 800)
}
}
