% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_lastonly.R
\name{geom_text_lastonly}
\alias{geom_text_lastonly}
\title{Label the last point(s) on a plot. Code mostly copied fromgeom_text source.}
\usage{
geom_text_lastonly(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = NULL,
  ...,
  parse = FALSE,
  nudge_x = 0.25,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Label the last point(s) on a plot. Code mostly copied fromgeom_text source.
}
