% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\name{get_cmapplot_global}
\alias{get_cmapplot_global}
\title{Get a value from the cmapplot_globals environment}
\usage{
get_cmapplot_global(name)
}
\description{
Get a value from the cmapplot_globals environment
}
