% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot_globals.R
\name{get_cmapplot_globals}
\alias{get_cmapplot_globals}
\alias{cmapplot_globals}
\alias{get_cmapplot_global}
\alias{set_cmapplot_global}
\title{The cmapplot_globals environment}
\usage{
get_cmapplot_globals()

get_cmapplot_global(...)

set_cmapplot_global(value, ..., quietly = FALSE)
}
\arguments{
\item{...}{The path to the variable within \code{cmapplot_globals} to be
get/set. The function willparse \code{$}, or recursive list elements can be
split over multiple arguments (e.g. \code{"font$strong$family"} is
equivalent to \code{"font", "strong", "family"}).}

\item{value}{the value to be set}

\item{quietly}{suppress confirmatory messages}
}
\description{
The \code{cmapplot_globals} environment contains a list of predefined
variables for use by the cmapplot package and its users. It includes commonly
used colors, font and font size specifications, and a list of constants which
aid in drawing cmap-themed plots. It cannot be accessed directly, but the
helper functions described here provide the user access if needed.
}
\section{Functions}{
\itemize{
\item \code{get_cmapplot_globals}: Get the entire environment as a list.

\item \code{get_cmapplot_global}: Get a specific global value

\item \code{set_cmapplot_global}: Set a specific global value
}}

\section{Plot Constants}{
 The primary portion of these global variables of
 interest to the user is \code{cmapplot_globals$consts}, a list of default
 constants that set certain plot aesthetics. Units of all plot constants are
 "bigpts": 1/72 of an inch. Most plot constants are invoked (and can be
 overridden) in \code{\link{finalize_plot}}: these are marked below with an
 \strong{F}. Some are used/can be overridden in \code{\link{theme_cmap}}:
 these are marked with \strong{T}.

 \itemize{ \item \code{lwd_strongline}: This stronger-width line is drawn
 vertically or horizontally with the \code{hline, vline} args of
 \code{theme_cmap()}. \strong{(T)} \item \code{lwd_gridline}: This
 thinner-width line is drawn vertically or horizontally with the
 \code{gridlines, axislines} args of \code{theme_cmap()}. \strong{(T)} \item
 \code{lwd_plotline}: The width of any lines drawn by geoms in the plot (e.g.
 \code{geom_line}) but not explicitly sized by the geom's aesthetic.
 Implemented by \code{finalize_plot} or by \code{apply_cmap_default_aes} but
 not overridable in either context. (Modify by setting the size explicitly in
 the geom, but see \code{gg_lwd_convert} first.) \item \code{lwd_topline}:
 The width of the line above the plot. \strong{(F)} \item
 \code{length_ticks}: The length of the axis ticks (if shown). \strong{(T)}
 \item \code{margin_topline_t}: The margin between the top edge of the image
 and the top line. \strong{(F)} \item \code{margin_title_t}: The margin
 between the top line and the title. \strong{(F)} \item
 \code{margin_title_b}: The margin between the title and the caption when
 both are drawn in the sidebar. \strong{(F)} \item \code{margin_caption_b}:
 The margin between the bottom of the caption and the bottom edge of the
 image. \strong{(F)} \item \code{margin_legend_t}: The margin between the top
 line and the plot box (i.e., the top of the legend). \strong{(F)} \item
 \code{margin_legend_i}: The margin between legends (this only applies in
 plots with two or more legends and does not affect legend spacing on plots
 with single legends that have multiple rows). \strong{(T, F)} \item
 \code{margin_legend_b}: The margin between the bottom of the legend and the
 rest of the plot. \strong{(T, F)} \item \code{margin_plot_b}: The margin
 between the bottom of the plot and the bottom edge of the image (or top of
 caption). \strong{(F)} \item \code{margin_sidebar_l}: The margin between the
 left edge of the image and the title and caption, when the sidebar exists.
 Deducted from \code{title_width}. \strong{(F)} \item \code{margin_plot_l}:
 The margin between the left edge of the plot and the sidebar. \strong{(F)}
 \item \code{margin_plot_r}: The margin between the right edge of the plot
 and the edge of the image. \strong{(F)} \item \code{margin_panel_r}: Padding
 between the plot and its right-hand drawing extent. Override this based on
 space needed for x axis labels. \strong{(T)} \item \code{leading_title}:
 Text leading for Title text. \strong{(F)} \item \code{leading_caption}: Text
 leading for Caption text. \strong{(F)} }
}

\examples{

# These are the same:
get_cmapplot_global("consts$lwd_gridline")
get_cmapplot_global("consts", "lwd_gridline")


# Globals can be modified if needed
set_cmapplot_global(5, "consts$lwd_gridline")
get_cmapplot_global("consts$lwd_gridline")

}
