% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\name{gg_size_convert}
\alias{gg_size_convert}
\title{Helper function to calculate correct size for ggplot line widths.}
\usage{
gg_size_convert(value, unit = "bigpts", use = c("line", "text"))
}
\arguments{
\item{value}{Numeric, the value to be converted.}

\item{type}{Char, the unit of the value to be converted. The type can be any
of the units accepted by \code{grid::unit()}, including "bigpt", "pt",
"mm", and "in". Default is \code{"bigpt"}.}
}
\value{
A unitless value that is the size in mm. For reasons not entirely clear,
  line values recieve an extra modification equivalent to\code{ggplot2::.stroke /
  ggplot2::.pt}.
}
\description{
Helper function to calculate correct size for ggplot line widths.
}
\seealso{
grid's \code{\link[grid]{unit}}, ggplot2's
  \code{\link[ggplot2]{.pt}}, and
  \url{https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size}
}
