% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grp_over_time}
\alias{grp_over_time}
\title{Gross Regional Product by cluster, 2007-17}
\format{A tibble. 121 rows and 5 variables:
\describe{
   \item{cluster}{Chr. The name of the cluster}
   \item{category}{Factor. "goods-producing" or "services"}
   \item{assessment}{Factor. "Trailing", "Mixed", or "Leading"}
   \item{year}{Double. The year of the data}
   \item{realgrp}{Double. The real gross regional product of the cluster in year `year`.
   Not exactly sure on the inflation year but I believe it is 2012}
}}
\source{
CMAP traded clusters report
}
\usage{
grp_over_time
}
\description{
A test dataset containing real GRP data for the CMAP region.
}
\examples{
# a time-series line chart
ggplot(grp_over_time, aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  theme_cmap()

}
\keyword{datasets}
