% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R, R/integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{A function factory for getting integer x- and y-axis values.}
\source{
\url{https://joshuacook.netlify.app/post/integer-values-ggplot-axis/}
}
\usage{
integer_breaks(n = 5, ...)

integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Numeric, desired number of breaks.}

\item{...}{other arguments passed on to \code{\link[base]{pretty}}}
}
\description{
This function can be supplied as the value for the \code{breaks} argument in
the \code{scale_*_continuous} ggplot elements. It forces labels to be
displayed as integers with even spacing between them, which is particularly
important for time series with years on the X or Y axis. The code was
developed by Joshua Cook.

Where n = desired number of ticks. Function uses \code{floor(pretty())} to
generate good breaks for the x or y axis of a ggplot. Borrowed with respect
from \url{https://joshuacook.netlify.com/post/integer-values-ggplot-axis/}
}
\examples{
# Standard implementation
ggplot(grp_over_time, aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  scale_x_continuous(breaks = integer_breaks())

# Adjusted to add a total of 11 intervals
ggplot(grp_over_time, aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  scale_x_continuous(breaks = integer_breaks(n = 11))


ggplot(data = dplyr::filter(grp_over_time, category == "Goods-Producing"),
mapping = aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  scale_x_continuous("Year", breaks = integer_breaks(n = 4)) +
  theme_minimal()

}
