% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_function.R
\name{make_highlight_palette}
\alias{make_highlight_palette}
\title{Highlight palette prep function}
\usage{
make_highlight_palette(variable, highlight, color = "#009ccc")
}
\description{
Create named list to serve as palette input for following fill/color highlight functions
}
