% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{percentile_wages}
\alias{percentile_wages}
\title{Wage percentiles by cluster}
\format{
A tibble. 45 rows and 3 variables:
\describe{
   \item{cluster}{Chr. The name of the cluster}
   \item{percentile}{Double. The percentile wage being reported}
   \item{wage}{Double. The wage. I believe 2017 data.}
}
}
\source{
CMAP traded clusters report
}
\usage{
percentile_wages
}
\description{
A test dataset containing the 10th, 25th, 50th, 75th, and 90th percentile wage by cluster in the CMAP region.
}
\examples{
# a non-time-series line chart
ggplot(percentile_wages, aes(x = percentile, y = wage, color = cluster)) +
  geom_line()

}
\keyword{datasets}
