% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_and_laborforce_by_age}
\alias{pop_and_laborforce_by_age}
\title{Population and Labor Force by Age}
\format{
A tibble. 12 rows and 4 variables:
\describe{
   \item{variable}{Chr. Indicates the meaning of the data stored in `value`. "population" or "laborforce".}
   \item{year}{Factor. 2010 or 2017.}
   \item{age}{Chr. The age bucket. Either 16-24, 25-54, or 55+.}
   \item{value}{Double. The value indicated by the other variables.}
}
}
\source{
CMAP traded clusters report
}
\usage{
pop_and_laborforce_by_age
}
\description{
A test dataset containing percentage breakdowns of the population and labor force by various age buckets in 2010 and 2017.
}
\examples{
# a grouped and stacked bar chart (via `interaction()`)
ggplot(pop_and_laborforce_by_age, aes(x = interaction(year, variable), y = value, fill = age)) +
  geom_col(position = position_stack(reverse = TRUE))

}
\keyword{datasets}
