% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\name{set_cmapplot_global}
\alias{set_cmapplot_global}
\title{Set a value in the cmapplot_globals environment}
\usage{
set_cmapplot_global(name, value)
}
\description{
Set a value in the cmapplot_globals environment
}
