% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traded_emp_by_race}
\alias{traded_emp_by_race}
\title{Traded employment by race}
\format{A tibble. 12 rows and 4 variables:
\describe{
   \item{Race}{Chr. White, Asian, Hispanic, Other, Black, or Regional Average.}
   \item{variable}{Chr. SpecializedTraded, UnspecializedTraded, or Total.
   Total is a sum of SpecializedTraded and UnspecializedTraded. The invisible remainder (e.g. `1-Total` or
   `1-(SpecializedTraded+UnspecializedTraded)`) is the percentage employed in local clusters.}
   \item{value}{Double. The value indicated by the other variables.}
}}
\source{
CMAP traded clusters report
}
\usage{
traded_emp_by_race
}
\description{
A test dataset containing the percentage breakdowns of the working population employed in traded clusters, by race.
}
\examples{
# a stacked bar chart
\dontshow{library(dplyr)}
filter(traded_emp_by_race, variable \%in\% c("SpecializedTraded", "UnspecializedTraded")) \%>\%
  ggplot(aes(x = reorder(Race, -value), y = value, fill = variable)) +
  geom_col(position = position_stack(reverse = TRUE)) +
  scale_y_continuous(labels = scales::percent)

}
\keyword{datasets}
