% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transit_ridership}
\alias{transit_ridership}
\title{Transit ridership in the Chicago region, 1980-2024}
\format{
A tibble. 225 rows and 3 variables
\describe{
   \item{year}{Double. Year of data}
   \item{system}{Char. Name of system (includes CTA bus, CTA rail, Metra, Pace, and Pace ADA)}
   \item{ridership}{Double. Annual unlinked passenger trips in millions}
}
}
\source{
Regional Transportation Authority \url{http://www.rtams.org/rtams/systemRidership.jsp}
}
\usage{
transit_ridership
}
\description{
A test dataset containing 1980-2019 transit ridership for the three service
boards that provide transit in Northeastern Illinois.
}
\examples{
# A line graph
ggplot(transit_ridership,aes(x = year,y=ridership,group=system,color=system)) +
  geom_line(na.rm=TRUE)


}
\keyword{datasets}
