% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_recessions.R
\name{update_recessions}
\alias{update_recessions}
\title{Update recessions table}
\source{
\url{https://data.nber.org/data/cycles/business_cycle_dates.json}
}
\usage{
update_recessions(url = NULL, quietly = FALSE)
}
\arguments{
\item{url}{Char, the web location of the NBER machine-readable CSV file. The
default, \code{NULL}, uses the most recently identified URL known to the
package development team, which appears to be the most stable location for
updates over time.}

\item{quietly}{Logical, suppresses messages produced by
\code{utils::download.file}.}
}
\value{
A data frame with the following variables: \itemize{ \item
 \code{start_char, end_char}: Chr. Easily readable labels for the beginning
 and end of the recession. \item \code{start_date, end_date}: Date. Dates
 expressed in R datetime format, using the first day of the specified month.
 \item \code{ongoing}: Logical. Whether or not the recession is ongoing as of
 the latest available NBER data. }
}
\description{
The cmapplot package contains an internal dataset \code{recessions} of all
recessions in American history as recorded by the National Bureau of Economic
Research (NBER). However, users may need to replace the built-in data, such as
in the event of new recessions and/or changes to the NBER consensus on
recession dates. This function fetches and interprets this data from the NBER
website.
}
\examples{
recessions <- update_recessions()

# package maintainers can update the internal dataset from within
# package by running the following code:
\dontrun{
 recessions <- update_recessions()
 usethis::use_data(recessions, pandemics, internal = TRUE, overwrite = TRUE)
}

}
