% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicle_ownership}
\alias{vehicle_ownership}
\title{Vehicle ownership in the CMAP seven county region}
\format{
A tibble. 40 rows and 3 variables
\describe{
   \item{county}{Char. Name of county}
   \item{number_of_veh}{Char. Number of vehicles owned by household}
   \item{pct}{Numeric. Share of households with the given number of vehicles (values between 0 and 1)}
}
}
\source{
CMAP Travel Inventory Survey Data Summary  \url{https://www.cmap.illinois.gov/documents/10180/77659/Travel+Inventory+Survey+Data+Summary_weighted_V2.pdf/d4b33cdd-1c44-4322-b32f-2f54b85207cb}
}
\usage{
vehicle_ownership
}
\description{
A test dataset containing vehicle ownership rates in the seven county region
of northeastern Illinois.
}
\examples{
# A stacked bar chart
ggplot(vehicle_ownership,
       aes(x = county, y = pct, fill = number_of_veh)) +
   geom_bar(position = position_stack(), stat = "identity")

}
\keyword{datasets}
