% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_continuous.R
\name{viz_gradient}
\alias{viz_gradient}
\title{Visualize CMAP color gradients}
\usage{
viz_gradient(pal, ttl = deparse(substitute(pal)))
}
\arguments{
\item{pal}{= select from cmap_gradients list}

\item{ttl}{= display title (optional)}
}
\description{
Displays 25 interpolated colors from the cmap continuous palettes.
Modeled after viz_palette from the \href{https://github.com/ropenscilabs/ochRe}{ochRe package}
}
\examples{
viz_gradient(cmap_gradients$multi_orange_red)

}
