% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_discrete.R
\name{viz_palette}
\alias{viz_palette}
\title{Print palette for reference}
\usage{
viz_palette(pal, ttl = deparse(substitute(pal)), num = length(pal))
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colours representing a palette}

\item{ttl}{character, title to be displayed (the name of the palette)}

\item{num}{numeric, the number of colours to display}
}
\description{
Directly from the \href{https://github.com/ropenscilabs/ochRe}{ochRe package}
}
\examples{
viz_palette(cmap_palettes$mobility)

}
