% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_handling.R
\name{abbr_years_cont}
\alias{abbr_years_cont}
\alias{abbr_years_date}
\title{Axis handling helper functions}
\usage{
abbr_years_cont(breaks, full_by_pos = c(1), full_by_year = NULL)

abbr_years_date(breaks, full_by_pos = c(1), full_by_year = NULL)
}
\arguments{
\item{breaks}{Vector, the input break values to be abbreviated. This will be
automatically supplied to the function by setting \code{labels =
abbr_years_cont} or \code{labels = abbr_years_date} inside the
\code{scale_*_continuous} argument (for \code{abbr_years_cont()} or inside
the \code{scale_*_date} argument (for \code{abbr_years_date}).}

\item{full_by_pos}{Vector of integers, the position of breaks that should not
be abbreviated. This defaults to \code{c(1)}, which retains the original
first label and abbreviates subsequent ones. If all breaks should be
abbreviated, this can be set to NULL.}

\item{full_by_year}{Vector of integers, the value of breaks that should not be
abbreviated. Defaults to NULL.}
}
\description{
`abbr_years_cont()` and `abbr_years_date()` are helper functions that allow
users to abbreviate year labels to their two-digit representation (e.g., 2008
to '08), but not abbreviating any specified breaks.
}
\section{Functions}{
\itemize{
\item \code{abbr_years_cont}: For continuous scales

\item \code{abbr_years_date}: For date scales
}}

\examples{
\dontrun{
# Default implementation - this will abbreviate all labels except the first
ggplot(transit_ridership \%>\% filter(year >= 2000),
       aes(x = year, y = ridership, color = system)) +
  geom_line() +
  scale_x_continuous(labels = abbr_years_cont)

# If customizations are desired, the function should be input using the format
# displayed below: \code{function(i) abbr_years_cont(i,full_by_pos = *,
# full_by_year = *)}.

# This example shows the use of \code{full_by_year} to maintain full labels on
# the first entry (the default) and add a full label to the year 2000, the first
# of the 21st century.
ggplot(transit_ridership,
       aes(x = year, y = ridership, color = system)) +
  geom_line() +
  scale_x_continuous(labels = function(i) abbr_years_cont(i,full_by_year = c(2000)))

# You can also remove the default maintenance of the first label and only
# specify specific years.
ggplot(grp_over_time, aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  scale_x_continuous(breaks = scales::breaks_pretty(5),
                     labels = function(i) abbr_years_cont(i,full_by_pos = NULL,
                                                          full_by_year = c(2018))
  )
}
}
