% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_continuous.R
\name{cmap_fill_continuous}
\alias{cmap_fill_continuous}
\alias{cmap_color_continuous}
\alias{cmap_colour_continuous}
\title{Apply continuous CMAP palettes (gradients) to ggplot2 aesthetics}
\usage{
cmap_fill_continuous(palette = "blues", reverse = FALSE, middle = 0, ...)

cmap_color_continuous(palette = "blues", reverse = FALSE, middle = 0, ...)

cmap_colour_continuous(palette = "blues", reverse = FALSE, middle = 0, ...)
}
\arguments{
\item{palette}{String; Choose from 'cmap_gradients' list}

\item{reverse}{Logical; Reverse color order?}

\item{middle}{Numeric; Sets midpoint for diverging color palettes. Default =
0.}

\item{...}{Additional parameters passed on to the scale type}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or
color). On diverging palettes, a midpoint can be manually adjusted (defaults
to 0). See \code{\link{cmap_gradients}} for a listing of available gradients.
}
\section{Functions}{
\itemize{
\item \code{cmap_fill_continuous()}: for fill aesthetic

\item \code{cmap_color_continuous()}: for color aesthetic

\item \code{cmap_colour_continuous()}: for color aesthetic

}}
\examples{
ggplot(dplyr::filter(grp_over_time, cluster=="Biopharmaceuticals"),
       aes(x = year, y = realgrp, color = realgrp)) +
    geom_line() +
    cmap_color_continuous(palette = "red_purple")

}
