% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_continuous.R
\name{cmap_fill_continuous}
\alias{cmap_fill_continuous}
\alias{cmap_color_continuous}
\alias{cmap_colour_continuous}
\title{Apply continuous CMAP palettes to ggplot2 aesthetics}
\usage{
cmap_fill_continuous(palette = "seq_reds", reverse = FALSE)

cmap_color_continuous(palette = "seq_reds", reverse = FALSE)

cmap_colour_continuous(palette = "seq_reds", reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'cmap_gradients' list}

\item{reverse}{Logical; reverse color order?}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or color)
}
\examples{
library(dplyr)
grp_over_time \%>\%
  filter(cluster=="Biopharmaceuticals") \%>\%
  ggplot(aes(x = year, y = realgrp, color = realgrp)) +
    geom_line() +
    cmap_color_continuous(palette = "multi_red_purple")

}
