% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_function.R
\name{cmap_fill_highlight}
\alias{cmap_fill_highlight}
\alias{cmap_color_highlight}
\alias{cmap_colour_highlight}
\title{Highlight one discrete group in a comparison graph}
\usage{
cmap_fill_highlight(variable, highlight, color = "#009ccc")

cmap_color_highlight(variable, highlight, color = "#009ccc")

cmap_colour_highlight(variable, highlight, color = "#009ccc")
}
\arguments{
\item{variable}{group vector, same as what is in your color/fill aesthetic. Include as 'data$variable' format}

\item{highlight}{name of group of interest, as character string}

\item{color}{(optional) Specify the highlighted color as a hexcode string. Default is #009ccc (blue)}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or color).
Specify the unique factor name of the group you're highlighting (as a case-sensitive string), and the vector it's found in.
}
\examples{
ggplot(dplyr::filter(transit_ridership, year=="2019"),
       aes(x = system, y = ridership, fill = system)) +
   geom_col() +
   cmap_fill_highlight(transit_ridership$system, "metra")

}
