% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_palette.R
\name{cmap_fill_race}
\alias{cmap_fill_race}
\alias{cmap_color_race}
\alias{cmap_colour_race}
\title{Apply official CMAP race/ethnicity chart colors to ggplot2}
\usage{
cmap_fill_race(white, african_american, hispanic, asian, other)

cmap_color_race(white, african_american, hispanic, asian, other)

cmap_colour_race(white, african_american, hispanic, asian, other)
}
\arguments{
\item{white}{White/Caucasian}

\item{african_american}{Black/African American}

\item{hispanic}{Hispanic/Latino}

\item{asian}{Asian}

\item{other}{Other}
}
\description{
Pick the function depending on the aesthetic of your ggplot object (fill or color).
Specify your dataset's unique race factor names (as case-sensitive strings) in the arguments.
All categories are optional in case your dataset does not have all of them.
}
\examples{
library(dplyr)
traded_emp_by_race \%>\%
    filter(Race!="Regional average" & Race!="Other") \%>\%
    ggplot(.) +
    geom_col(aes(x = variable, y = value, fill = Race)) +
    cmap_fill_race(white = "White", african_american = "Black",
                   hispanic = "Hispanic", asian = "Asian")

}
