% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_geom_defaults.R
\name{cmap_geom_defaults}
\alias{cmap_geom_defaults}
\alias{fetch_cmap_geom_defaults}
\alias{set_cmap_geom_defaults}
\title{Fetch and set aesthetic defaults}
\usage{
fetch_cmap_geom_defaults()

set_cmap_geom_defaults(values = NULL, quietly = FALSE)
}
\arguments{
\item{values}{A list of lists representing the aesthetic defaults to set for
the impacted geoms. Leave null, the default, to use the CMAP-preferred
values. Use this field to input the output of \code{fetch_cmap_geom_defaults}
in order to reset to ggplot defaults.}

\item{quietly}{Suppress confirmatory message}
}
\description{
These functions allow for fetching and setting default aesthetic values for
certain ggplot2 geoms. This is necessary for geoms to be "themed" to CMAP
style standards, because (at least at the moment) setting geom aesthetic
defaults on a plot-by-plot basis (such as with \code{ggplot2::theme}) is not
possible. The geoms impacted and their preferred settings are hardwired into
the functions, and at the moment include \code{Line}, \code{Text}, and the
text-based custom \code{cmapplot} geoms \code{TextLast} and
\code{RecessionsText}.
}
\details{
These functions are employed implictly within \code{\link{finalize_plot}} to
apply preferred aesthetic defaults to final outputs. They are only necessary
to use explicitly if you would like plots to use these defaults
pre-\code{finalize}.

CAUTION: Running \code{set_cmap_geom_defaults} will set defaults for any
ggplot drawn in the current session, including those without
\code{theme_cmap}. If this is not desired, you must re-run the function with
old values (see examples) or unload and reload the ggplot2 package when
you're done.
}
\section{Functions}{
\itemize{
\item \code{fetch_cmap_geom_defaults}: Fetch geom defaults

\item \code{set_cmap_geom_defaults}: Set geom defaults, either to CMAP-preferred or
user-inputed values
}}

\examples{
\dontrun{
g <- ggplot(mtcars, aes(x = mpg, y = disp, label = rownames(mtcars))) +
  geom_line() +
  geom_text() +
  geom_text_lastonly() +
  theme_minimal() +
  coord_cartesian(clip = "off")

# print normally
g

# before overwriting defaults, save the current ones
old <- fetch_cmap_geom_defaults()

# overwite defaults
set_cmap_geom_defaults()
g

# reset to original defaults
set_cmap_geom_defaults(old)
g

# finalize alters the defaults implicitly, but then resets them automatically.
finalize_plot(g)

# you can also use `finalize` without these modifications
finalize_plot(g, use_cmap_aes = FALSE)
}

}
