% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_continuous.R
\docType{data}
\name{cmap_gradients}
\alias{cmap_gradients}
\alias{viz_gradient}
\title{CMAP continuous color palettes (gradients)}
\format{
An object of class \code{list} of length 23.
}
\usage{
cmap_gradients

viz_gradient(pal, ttl = deparse(substitute(pal)))
}
\arguments{
\item{pal}{= select from cmap_gradients list}

\item{ttl}{= display title (optional)}
}
\description{
Sequential and divergent gradients extracted from the CMAP color palette.
}
\section{Functions}{
\itemize{
\item \code{viz_gradient}: Display CMAP continuous palettes, interpolating
additional colors as needed. Modeled after viz_palette from the
\href{https://github.com/ropenscilabs/ochRe}{ochRe package}
}}

\examples{
# Get names of available continuous palettes.
# (Call viz_palette("name_of_palette") to preview one.)
names(cmap_gradients)

# Run the following function to visualize *all* continuous palettes
purrr::walk2(cmap_gradients, names(cmap_gradients), viz_gradient)

}
\keyword{datasets}
