% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_continuous.R
\docType{data}
\name{cmap_gradients}
\alias{cmap_gradients}
\title{CMAP continuous color palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
cmap_gradients
}
\description{
A sample of comms-approved colors arranged into palettes
}
\examples{
viz_gradient(cmap_gradients$red_purple)

}
\keyword{datasets}
