% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_continuous.R
\name{cmap_pal_continuous}
\alias{cmap_pal_continuous}
\title{Continuous palette prep function}
\usage{
cmap_pal_continuous(palette = "seq_reds", reverse = FALSE)
}
\description{
Continuous palette prep function
}
