% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_discrete.R
\docType{data}
\name{cmap_palettes}
\alias{cmap_palettes}
\alias{viz_palette}
\title{CMAP discrete color palettes}
\format{
An object of class \code{list} of length 31.
}
\usage{
cmap_palettes

viz_palette(pal, ttl = deparse(substitute(pal)), num = length(pal))
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colors representing a palette}

\item{ttl}{character, title to be displayed (the name of the palette)}

\item{num}{numeric, the number of colors to display}
}
\description{
A selection of discrete color palettes from the CMAP color palette. These
include mixed color palettes and discrete versions of the gradients defined
in \code{link{cmap_fill_continuous}}.
}
\section{Functions}{
\itemize{
\item \code{viz_palette}: Display CMAP palettes. Borrowed with respect from
the \href{https://github.com/ropenscilabs/ochRe}{ochRe package}
}}

\examples{
# Get names of available discrete palettes.
# (Call viz_palette(cmap_palettes$name_of_palette) to preview one.)
names(cmap_palettes)

# Run the following function to visualize *all* discrete palettes
purrr::walk2(cmap_palettes, names(cmap_palettes), viz_palette)

}
\keyword{datasets}
