% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_discrete.R
\docType{data}
\name{cmap_palettes}
\alias{cmap_palettes}
\title{CMAP discrete color palettes}
\format{An object of class \code{list} of length 8.}
\usage{
cmap_palettes
}
\description{
A sample of comms-approved colors arranged into palettes
}
\examples{
viz_palette(cmap_palettes$mobility)

ggplot(pop_and_laborforce_by_age, aes(x = variable, y = value, fill = age)) +
   geom_col(position = position_stack(reverse = TRUE)) +
   facet_wrap(~year) +
   cmap_fill_discrete(palette = "community")

}
\keyword{datasets}
