% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\docType{package}
\name{cmapplot}
\alias{cmapplot}
\alias{cmapplot-package}
\title{cmapplot}
\description{
This package contains extra palettes, themes and geoms for \pkg{ggplot2}.
}
\details{
Please see the help pages listed below:

\itemize{
  \item \code{\link{geom_text_lastonly}}
  \item \code{\link{theme_cmap}}
}

Please report issues and suggest improvements at Github:

\url{https://github.com/CMAP-REPOS/cmapplot}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmap-repos.github.io/cmapplot}
}

}
\author{
\strong{Maintainer}: Matthew Stern \email{mstern@cmap.illinois.gov} (\href{https://orcid.org/0000-0001-9119-2121}{ORCID})

Authors:
\itemize{
  \item Noel Peterson \email{npeterson@cmap.illinois.gov} (\href{https://orcid.org/0000-0003-1290-2362}{ORCID})
  \item Daniel Comeaux \email{dcomeaux@cmap.illinois.gov} (\href{https://orcid.org/0000-0002-8412-9092}{ORCID})
}

Other contributors:
\itemize{
  \item Chicago Metropolitan Agency for Planning [copyright holder]
}

}
\keyword{internal}
