% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\docType{package}
\name{cmapplot}
\alias{cmapplot-package}
\alias{cmapplot}
\title{cmapplot}
\description{
This package contains extra palettes, themes and geoms for \pkg{ggplot2},
based on Chicago Metropolitan Agency for Planning (CMAP) design guidelines.
}
\details{
Detailed documentation can be viewed at
\url{https://cmap-repos.github.io/cmapplot}.

Please report issues and suggest improvements at
\url{https://github.com/CMAP-REPOS/cmapplot/issues}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmap-repos.github.io/cmapplot}
  \item \url{https://github.com/CMAP-REPOS/cmapplot}
  \item Report bugs at \url{https://github.com/CMAP-REPOS/cmapplot/issues}
}

}
\author{
\strong{Maintainer}: Sean Connelly \email{sconnelly@cmap.illinois.gov}

Authors:
\itemize{
  \item Daniel Comeaux (\href{https://orcid.org/0000-0002-8412-9092}{ORCID})
  \item Sarah Buchhorn
  \item Martin Menninger \email{mmenninger@cmap.illinois.gov}
  \item Noel Peterson (\href{https://orcid.org/0000-0003-1290-2362}{ORCID})
  \item Greta Ritzenthaler
  \item Matthew Stern (\href{https://orcid.org/0000-0001-9119-2121}{ORCID})
  \item David Wells \email{dwells@cmap.illinois.gov}
}

Other contributors:
\itemize{
  \item Chicago Metropolitan Agency for Planning [copyright holder, funder]
}

}
\keyword{internal}
