% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmapplot.R
\docType{package}
\name{cmapplot}
\alias{cmapplot}
\alias{cmapplot-package}
\title{cmapplot}
\description{
This package contains extra palettes, themes and geoms for \pkg{ggplot2},
based on Chicago Metropolitan Agency for Planning (CMAP) design guidelines.
}
\details{
Detailed documentation can be viewed at
\url{https://cmap-repos.github.io/cmapplot}.

Please report issues and suggest improvements at
\url{https://github.com/CMAP-REPOS/cmapplot/issues}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmap-repos.github.io/cmapplot}
  \item \url{https://github.com/CMAP-REPOS/cmapplot}
}

}
\author{
\strong{Maintainer}: Matthew Stern \email{mstern@cmap.illinois.gov} (\href{https://orcid.org/0000-0001-9119-2121}{ORCID})

Authors:
\itemize{
  \item Sarah Buchhorn \email{sbuchhorn@cmap.illinois.gov}
  \item Daniel Comeaux \email{dcomeaux@cmap.illinois.gov} (\href{https://orcid.org/0000-0002-8412-9092}{ORCID})
  \item Martin Menninger \email{mmenninger@cmap.illinois.gov}
  \item Noel Peterson \email{npeterson@cmap.illinois.gov} (\href{https://orcid.org/0000-0003-1290-2362}{ORCID})
  \item Greta Ritzenthaler \email{gritzenthaler@cmap.illinois.gov}
}

Other contributors:
\itemize{
  \item Chicago Metropolitan Agency for Planning [copyright holder]
}

}
\keyword{internal}
