% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_recessions.R, R/geom_text_lastonly.R
\docType{data}
\name{GeomRecessions}
\alias{GeomRecessions}
\alias{customproto}
\alias{GeomTextLast}
\alias{GeomPointLast}
\title{Custom ggproto classes}
\description{
The \code{cmapplot} package contains a few custom ggproto objects. For the
most part, these are slightly tweaked versions of ggplot2's default proto
objects. For more information about these, see
\code{\link[ggplot2]{ggplot2-ggproto}}.
}
\keyword{datasets}
