% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_plot2.R
\name{finalize_plot2}
\alias{finalize_plot2}
\title{Arrange and save CMAP ggplot chart}
\usage{
finalize_plot2(input_plot = ggplot2::last_plot(), title = "Title here",
 caption = "Caption here", mode = c("plot", "newwindow", "object", "png",
 "tiff", "jpeg", "bmp", "svg", "ps", "pdf"), width = 7, height = 4,
 title_width = 1.5, filepath = NULL, plot_margins = c(5,5,5,10), topline = 2,
 topline_margin = 5, text_margin_left = 2, text_margin_top = 5,
 text_margin_mid = 10, white_canvas = FALSE)
}
\arguments{
\item{input_plot}{ggplot object, the variable name of the plot you have
created that you want to finalize. The default is
\code{ggplot2::last_plot()}, so unless specified the most recent plot will
be used.}

\item{title}{Char, the text you want to appear in the title block.}

\item{caption}{Char, the text you want to appear in the caption block.}

\item{mode}{Char, the action to be taken with the plot. Default is to view in
the window (using "plot"). Other options include "newwindow" (displays in a
new window), "object" (returns a grob object, but does not print), or one of
file extensions that can be saved (png, tiff, jpeg, bmp, svg, ps, and pdf).}

\item{width}{Numeric, the width in inches for the image, including the title.
Default = 7.}

\item{height}{Numeric, the height in inches for the image. Default = 4.}

\item{title_width}{Numeric, the width in inches for the title. Default = 1.5.}

\item{filepath}{Char, the filepath you want the plot to be saved to. You can
specify an extension to use, or an extension will be added if you specify it
in "mode".}

\item{plot_margins}{Vector of units, the margins around the elements of the
plot within the plot object. This requires a vector of 4 "unit" elements,
defining the margins clockwise starting from the top. The default is 5, 5,
5, and 10 big points  (1/72 of an inch) on the top, right, bottom, and left,
respectively. Inputs should be formatted as c(`top`,`right`,`bottom`,`left`)}

\item{topline}{Numeric, the width of the line above the title and graph in big
points. Default is 3 "big points."}

\item{topline_margin}{Numeric, the margin between the top line and the top of
the exported image. Default is 5 "big points."}

\item{text_margin_left}{Unit, the margin to left of title and caption text.
Default is 2 "big points."}

\item{text_margin_top}{Unit, the margin between the top line and title text.
Default is 5 "big points."}

\item{text_margin_mid}{Unit, the margin between the title and caption text.
Default is 10 "big points."}

\item{white_canvas}{Bool, whether the canvas for "plot" or "newwindow" should
be white outside the plot element. Default is FALSE, which returns a gray
canvas to make it easier to determine whether text is overflowing the
desired size.}
}
\description{
\code{finalize_plot2} will save a ggplot into a frame defined by  CMAP design
standards. It will align your title and caption to the left, add a horizontal
line on top, and make other adjustments. If instructed to do so, it will also
save the plot in one of four file formats (png, tiff, svg, and pdf) to a
specified location. This function will not apply CMAP design standards to the
plot itself: use with \code{theme_cmap()} for that.
}
\examples{
\dontrun{
econ_plot <-
  cluster_jobchange \%>\%
  ggplot(aes(x = reorder(name, jobchange), y = jobchange, fill = category)) +
  geom_col() +
  coord_flip() +
  theme_cmap(gridlines = "v", hline = 0) +
  scale_y_continuous(labels = scales::comma)

finalize_plot2(econ_plot,
               "Cluster-level employment changes in the Chicago MSA, 2001-17.",
               "Source: Chicago Metropolitan Agency for Planning analysis.",
               mode = "plot",
               height = 6,
               plot_margins = c(5,30,5,10))

transit_plot <- transit_ridership \%>\%
  mutate(system = case_when(
    system == "cta_bus" ~ "CTA (Bus)",
    system == "cta_rail" ~ "CTA (Rail)",
    system == "metra" ~ "Metra",
    system == "pace" ~ "Pace",
    system == "pace_ada" ~ "Paratransit"
  )) \%>\%
  ggplot(aes(x = year, y = ridership, color = system)) +
  geom_line() +
  theme_cmap(max_columns = 3)

finalize_plot2(transit_plot,
               "Transit ridership in the RTA region over time, 1980-2019
               (in millions).",
               "Source: Chicago Metropolitan Agency for Planning
               analysis of data from the Regional Transportation Authority.",
               mode="png",
               filepath = "foo",
               title_width = 2,
               plot_margins = c(5,10,5,10))
}
}
