% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_lastonly.R
\name{geom_text_lastonly}
\alias{geom_text_lastonly}
\title{Text (Last Only)}
\usage{
geom_text_lastonly(mapping = NULL, data = NULL, stat = "identity",
  position = NULL, parse = FALSE, nudge_x = 0.25, nudge_y = 0,
  check_overlap = FALSE, na.rm = FALSE, show.legend = FALSE, inherit.aes =
  TRUE, add_points = FALSE, text_aes = NULL, point_aes = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function. Cannot be jointy specified with
\code{nudge_x} or \code{nudge_y}.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels by.
Useful for offsetting text from points, particularly on discrete scales.
Cannot be jointy specified with \code{position}.}

\item{check_overlap}{If \code{TRUE}, text that overlaps previous text in the
same layer will not be plotted. \code{check_overlap} happens at draw time
and in the order of the data. Therefore data should be arranged by the
label column before calling \code{geom_text_lastonly()}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{add_points}{If \code{TRUE}, points will be added to the plot (for the
labeled data only). Default size=2, color will match line color.}

\item{text_aes, point_aes}{Named list, additional aesthetics to send to the
text and point geoms, respectively.}

\item{...}{additional aesthetics to send to BOTH the point and text geoms.
Note that if \code{add_points = FALSE}, additional parameters can be passed
to the text geom here, rather than in \code{text_aes}, without breaking.}
}
\description{
Label only the last point(s) on a plot. \code{geom_text_lastonly()} can be
used instead of \code{ggplot2::geom_text()} when only the last point(s)
should be labeled. This is accomplished by identifying the maximum value of
\code{x} in \code{data} and applying a filter to omit records where \code{x}
is less than the maximum.
}
\details{
Labels are automatically placed to the right of the final point, and may be
partially cut off by the plot limits, unless the \code{x} scale is expanded,
e.g. with \code{scale_x_continuous(expand=expand_scale(mult=0.10))}.

Code was mostly copied from the source of \code{ggplot2::geom_text()} and
\code{ggplot2::geom_point()}.
}
\examples{
df <- data.frame(year=2010:2020, value=runif(22), var=c(rep("A", 11), rep("B", 11)))

# Without points, label formatting or x-axis expansion
ggplot(df, aes(x=year, y=value, color=var)) +
  geom_line() +
  labs(title="Random lines") +
  scale_y_continuous("Percentage of absolutely nothing") +
  scale_x_continuous("Year") +
  geom_text_lastonly()

# With points, label formatting and x-axis expansion
ggplot(df, aes(x=year, y=value, color=var, label=sprintf("\%.1f\%\%", 100*var))) +
  geom_line() +
  labs(title="Random lines") +
  scale_y_continuous("Percentage of absolutely nothing", labels=scales::percent) +
  scale_x_continuous("Year", expand=expansion(mult=c(0.05, 0.10))) +
  geom_text_lastonly(add_points=TRUE, text_aes=list(fontface="bold"), point_aes=list(size=2.5))

}
