% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{clean up integer axis breaks}
\usage{
scale_x_continuous(breaks = integer_breaks(n = 5))
}
\arguments{
\item{n}{Numeric, desired number of breaks.}

\item{...}{other arguments passed on to \code{\link[base]{pretty}}}
}
\description{
Where n = desired number of ticks. Function uses \code{floor(pretty())} to
generate good breaks for the x or y axis of a ggplot. Borrowed with respect
from \url{https://joshuacook.netlify.com/post/integer-values-ggplot-axis/}
}
\examples{

ggplot(data = dplyr::filter(grp_over_time, category == "Goods-Producing"),
mapping = aes(x = year, y = realgrp, color = cluster)) +
  geom_line() +
  scale_x_continuous("Year", breaks = integer_breaks(n = 4)) +
  theme_minimal()

}
