% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peer_grp}
\alias{peer_grp}
\title{Gross Regional Product by region, with peers, 2001-17}
\format{
A tibble. 121 rows and 5 variables:
\describe{
   \item{Area}{Char. The name of the region}
   \item{year}{Double. The year of the data}
   \item{realgrp}{Double. The real gross regional product of the cluster in year `year`.
   Not exactly sure on the inflation year but I believe it is 2012}
}
}
\source{
CMAP traded clusters report
}
\usage{
peer_grp
}
\description{
A test dataset containing real GRP data for the CMAP and peer regions.
}
\examples{
# a time-series line chart
ggplot(grp_over_time, aes(x = year, y = realgrp, color = cluster)) +
  geom_line()

}
\keyword{datasets}
