% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recessions}
\alias{recessions}
\title{A list of all recessions in American history}
\format{A tibble. 33 rows and 4 variables:
\describe{
   \item{peak, trough}{Chr. Easily readable labels for the beginning and end of the recession}
   \item{recession_start, recession_end}{Double. Dates expressed as years, with decimels referring to months. Need to improve this.}
}}
\source{
Update this.
}
\usage{
recessions
}
\description{
An internal dataset containing a list of all recessions in American history, as recorded by
the National Bureau of Economic Research (NBER). As future recessions occur, this file will
need to be updated using the script below.
}
\keyword{datasets}
