% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_continuous.R, R/colors_discrete.R
\name{viz_gradient}
\alias{viz_gradient}
\alias{viz_palette}
\alias{cmap_palettes}
\alias{cmap_gradients}
\alias{cmap_colors}
\title{Visualize CMAP color gradients}
\usage{
viz_gradient(pal, ttl = NULL)

viz_palette(pal, ttl = NULL, num = NULL)
}
\arguments{
\item{pal}{character, name of a a cmapplot palette, or a vector of colors
representing a palette}

\item{ttl}{character, title to be displayed (the name of the palette)}

\item{num}{numeric, the number of colors to display}
}
\description{
The cmapplot package contains a many color palettes extracted from the
larger, official CMAP color palette. Helper functions allow the user to
inspect the various palettes before applying them to plots.
}
\details{
Palettes are stored in a tibble the \code{cmapplot_globals} environment. The
user can access this tibble with \code{\link{get_cmapplot_global}}. For more
information about available cmapplot color palettes and how to apply them,
see \code{vignette("colors")}.

These functions are modified with respect from the
\href{https://github.com/ropenscilabs/ochRe}{ochRe package})
}
\section{Functions}{
\itemize{
\item \code{viz_gradient}: interpolates the range of colors a sequential or
divergent palette offers when used on a continuous scale.

\item \code{viz_palette}: displays the colors of any cmapplot palette
}}

\examples{
# Vizualize a sequential or divergent palette with interpolation
viz_gradient("reds")

# Visualize a single palette as individual colors
viz_palette("prosperity")

# Get names of all available palettes.
as.data.frame(get_cmapplot_global("palettes")[1:2])

}
